% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.structure.plot.R
\name{Plots for Tests of Pathway Structure in iGraph}
\alias{Plots for Tests of Pathway Structure in iGraph}
\alias{pathway_perm_plot_density}
\alias{pathway_perm_plot_raw}
\title{Extensions to iGraph for Plotting Pathway Structure}
\usage{
pathway_perm_plot_raw(table_obj, main = NULL, sub = NULL,
  xlab = "up events", ylab = "down events", col = "black", cex = 1,
  pch = 19)

pathway_perm_plot_density(table_obj, main = NULL, xlab = "up - down events",
  ylab = "density")
}
\arguments{
\item{table_obj}{List. Output of \code{\link[permutation.structure]{pathway.structure.permuation}}}

\item{main, sub, xlab, ylab, col, cex, pch}{Arguments to passs on to \code{\link[plot]{graphics}} or \code{\link[plot.density]{stats}}}
}
\description{
Functions to plot outcomes for testing pathway structure with permutation analysis.
}
\examples{

#generate example graphs
library("igraph")
graph <- make_graph(unlist(lapply(letters, function(x) rep(x, 2)))[2:51], directed = TRUE)
plot(graph)

#Example states
test.structure(graph, "a", "c")
source <- sample(letters, 10)
target <- sample(letters, 10)
V(graph)$color <- ifelse(names(V(graph)) \%in\% source, "lightblue", "grey75")  
V(graph)$color <- ifelse(names(V(graph)) \%in\% target, "palevioletred", V(graph)$color)  
V(graph)$color <- ifelse(names(V(graph)) \%in\% intersect(source, target), "mediumpurple2", V(graph)$color)  
plot(graph, layout = layout.fruchterman.reingold, vertex.color= V(graph)$color, vertex.label.family = "mono", vertex.size = 10, vertex.label.color = "black", vertex.frame.color= "grey50", main = g$name)
matrix.structure(graph, source, target)
table(matrix.structure(graph, source, target))
perm_table <- permutation.structure(graph, source, target, letters)

#Plot Raw Output of Permutation Test
pathway_perm_plot_raw(perm_table)

#Plot Density Output of Permutation Test
pathway_perm_plot_density(perm_table)
}

