% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.structure.R
\name{Tests of Pathway Structure in iGraph}
\alias{Tests of Pathway Structure in iGraph}
\alias{pathway_test}
\title{Extensions to iGraph for Pathway Structure}
\usage{
pathway_test(graph, target_node, source_node, shortest.paths.in = NULL)
}
\arguments{
\item{graph}{An \code{\link[igraph]{igraph}} object. May be directed or weighted as long as a shortest path can be computed.}

\item{shortest.paths.in}{Defaults to NULL leading to computing the shortest paths from the input graph where necessary, these may be given as computed in advance (or passed from higher functions) to reduce computational redundancy.}

\item{target}{string: vector of target states for testing pathway stucture. Must be a single string for target_node. These are cross referenced against V(graph)$name.}

\item{source}{string: vector of source states for testing pathway stucture. Must be a single string for source_node. These are cross referenced against V(graph)$name.}

\item{universe}{string vector of potential nodes to be assigned target and\/or source states. This may be V(graph)$name, and subset thereof, or a larger pool of nodes to assign states for permutation analysis.}

\item{reps}{scalar numeric. Number of permutations to statistically test the structure of the network.}
}
\description{
Functions to compute the pathway structure of different states within a network with testing pathway structure with permutation analysis. Uses shortest paths to compute the structure between nodes of differing states. Shortest paths are computed in advance where possible to reduce computational redundancy.
}
\examples{

#generate example graphs
library("igraph")
g1 <- make_ring(10)
V(g1)$name <- letters[1:10]
g2 <- make_star(10)
V(g2)$name <- letters[1:10]

#test pathway structure between two points
pathway_test(g1, "a", "c")
pathway_test(g2, "a", "c")
}

